Modified the way to get right block device partition to test.
Before, get_blk_device_node return the biggest partition except boot partition on the device;
If the partition is rootfs partition and it also is the biggest one,
the test may format this partition which results in loss of rootfs in blk device.
With this patch, get_blk_device_node return the biggest partition if the partition is not boot or rootfs partition.
Only tested on MMC. Should work on other types of blk devices like usb, ata etc