ddt/usb_test_suite: testusb: Add do_test() function
The original handle_testdev() funtion is more suited for tests
to be run on multiple devices as threads and it doesn't return error code.

We add the do_test() function that runs tests on a single device and
bails out with an error code on a failed test case.

do_test() will only be invoked when there is a single test device
in the system or a specific test device is specified using command
line argument D.

Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>