LTP-DDT Test cases for multiple ethernet interfaces
This patch adds functional and stress tests for all ethernet interfaces
supported on a platform. They include tests to bring down and up the
interfaces all at the same time, one at a time, as well as ping tests
on these interfaces and combinations of ifdown/ifup on other eth interfaces
in parallel with ping on one interface.

Signed-off-by: Aparna Balasubramanian <aparnab@ti.com>

Changes for v3 - Renamed "stress" type to "ifupdown" so that it is not
                 confused with test scope

Changes for v2 - Changed the nested else by replacing with elif at:
                 > +else
                 > +  if [ "$p_sequence" = 'one' ]

Signed-off-by: Aparna Balasubramanian <aparnab@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>