Simplify multi_run_processes.sh
multi_run_processes is doing more than it should. This script is
meant to run multiple processes in parallel with different cpu
affinities and priorities. It should not make any assumptions
about what the processes are doing.

This patch also restores run_processes.sh to original content
where it prints processes stdout and stderr to the console
for logging purposes.

Signed-off-by: Carlos Hernandez <ceh@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>