LTP-DDT patch for usb eth adapter (USBHOST-ECM class) tests
This patch adds test cases to perform ping test on usb ethernet adapter
interfaces. It adds the supporting scripts to search for the interface
and also to find all active ethernet interfaces at beginning of
test so that they can be restored at the end of the test.

Version 2 - removed include st.log since that is not required and taken
care of in common files.

Signed-off-by: Aparna Balasubramanian <aparnab@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>