Create test partition if there is no any partitions on mmc or emmc
Since most of emmc cards on omap5 or J5 don't have any partition for testing, add this patch to create test partition to reduce the manual work.
This patch create one test partition only if there is no any partitions in the card.
This patch only add this logic for emmc for now. It could be added to mmc if needed.

Signed-off-by: Yan Liu <yan-liu@ti.com>